formatted files in accordance to betty linter

+++ 0x1A-hash_tables/5-hash_table_print.c
@@ -15,7 +15,7 @@ void hash_table_print(const hash_table_t *ht)
 		return;
 
 	printf("{");
-	for (i = 0; i < ht->size; i++)
+	for (i = 0; i < (int)ht->size; i++)
 	{
 		item = ht->array[i];
 		if (item == NULL)
